bunch of refactor
