+++ server_core/workflows/bugbounty/testing.py
@@ -1,4 +1,4 @@
-from typing import List, Dict, Any
+from typing import Dict, Any
 
 class FileUploadTestingFramework:
     """Specialized framework for file upload vulnerability testing"""
